Conversão do valor da quantidade de produtos a serem comprados para decimal, evitando inconsistencia na multiplicação de quantidade com o preço do produto.

+++ Ecommerce.Domain/Entities/TransactionEntity.cs
@@ -29,7 +29,7 @@ namespace Ecommerce.Domain.Entities
         
         public void CalculateTotalValue()
         {
-            TotalValue = TransactionProductsList.Sum(x => x.Quantity * x.Product.Price);
+            TotalValue = TransactionProductsList.Sum(x => (decimal)x.Quantity * x.Product.Price);
         }
 
         public void SetTransactionStatusToPaid()
